new: Compress response from PlaywrightCapture
Instead of dumping the response to JSON and storing that in redis, we
keep the python object, pickelize and compress it, and finally store it
as-is in redis. It reduces the RAM usage by 3.
The public method returning the response doesn't change its input and
output in order to keep the API compatible.